+++ rs/earthbucks_lib/src/tx_out_bn_map.rs
@@ -28,7 +28,7 @@ impl TxOutBnMap {
         name.split(':').nth(1).unwrap().parse().unwrap()
     }
 
-    pub fn add(&mut self, tx_id: &[u8; 32], tx_out_num: u32, tx_out: TxOut, block_num: u64) {
+    pub fn add(&mut self, tx_id: &[u8; 32], tx_out_num: u32, tx_out: TxOut, block_num: u32) {
         let name = Self::name_from_output(tx_id, tx_out_num);
         let tx_out_bn = TxOutBn {
             tx_out: tx_out.clone(),
@@ -51,7 +51,7 @@ impl TxOutBnMap {
         self.map.values().collect()
     }
 
-    pub fn add_tx_outputs(&mut self, tx: &Tx, block_num: u64) {
+    pub fn add_tx_outputs(&mut self, tx: &Tx, block_num: u32) {
         for (output_index, output) in tx.outputs.iter().enumerate() {
             self.add(&tx.id(), output_index as u32, output.clone(), block_num);
         }
@@ -86,7 +86,7 @@ mod tests {
             0, 0, 0,
         ];
         let output_index = 0;
-        let block_num: u64 = 0;
+        let block_num: u32 = 0;
         let tx_out_bn = TxOutBn {
             tx_out: tx_output.clone(),
             block_num,
